Consolidated reporting and validation to single file
Updated reporting to color code outputs
